TestCapacityScheduler: queue and refresh related tests moved to separated file
